don't rsc render on sub routes
