+++ templates/demo-store/src/components/global/Header.client.tsx
@@ -39,6 +39,8 @@ export function Header({title, menu}: {title: string; menu?: EnhancedMenu}) {
     closeDrawer: closeMenu,
   } = useDrawer();
 
+  console.log('Render header');
+
   return (
     <>
       <CartDrawer isOpen={isCartOpen} onClose={closeCart} />
