+++ templates/demo-store/src/routes/component/HeaderMenu.server.tsx
@@ -1,0 +1,96 @@
+import {useLocalization, useShopQuery, CacheLong, gql} from '@shopify/hydrogen';
+import type {Menu, Shop} from '@shopify/hydrogen/storefront-api-types';
+
+import {Header} from '~/components';
+import {parseMenu} from '~/lib/utils';
+
+const HEADER_MENU_HANDLE = 'main-menu';
+const FOOTER_MENU_HANDLE = 'footer';
+
+const SHOP_NAME_FALLBACK = 'Hydrogen';
+
+export default function HeaderMenu() {
+  const {shopName, headerMenu} = useLayoutQuery();
+  return <Header title={shopName} menu={headerMenu} />;
+}
+
+function useLayoutQuery() {
+  const {
+    language: {isoCode: languageCode},
+  } = useLocalization();
+
+  const {data} = useShopQuery<{
+    shop: Shop;
+    headerMenu: Menu;
+    footerMenu: Menu;
+  }>({
+    query: SHOP_QUERY,
+    variables: {
+      language: languageCode,
+      headerMenuHandle: HEADER_MENU_HANDLE,
+      footerMenuHandle: FOOTER_MENU_HANDLE,
+    },
+    cache: CacheLong(),
+    preload: '*',
+  });
+
+  const shopName = data ? data.shop.name : SHOP_NAME_FALLBACK;
+
+  /*
+    Modify specific links/routes (optional)
+    @see: https://shopify.dev/api/storefront/unstable/enums/MenuItemType
+    e.g here we map:
+      - /blogs/news -> /news
+      - /blog/news/blog-post -> /news/blog-post
+      - /collections/all -> /products
+  */
+  const customPrefixes = {BLOG: '', CATALOG: 'products'};
+
+  const headerMenu = data?.headerMenu
+    ? parseMenu(data.headerMenu, customPrefixes)
+    : undefined;
+
+  const footerMenu = data?.footerMenu
+    ? parseMenu(data.footerMenu, customPrefixes)
+    : undefined;
+
+  return {footerMenu, headerMenu, shopName};
+}
+
+const SHOP_QUERY = gql`
+  fragment MenuItem on MenuItem {
+    id
+    resourceId
+    tags
+    title
+    type
+    url
+  }
+  query layoutMenus(
+    $language: LanguageCode
+    $headerMenuHandle: String!
+    $footerMenuHandle: String!
+  ) @inContext(language: $language) {
+    shop {
+      name
+    }
+    headerMenu: menu(handle: $headerMenuHandle) {
+      id
+      items {
+        ...MenuItem
+        items {
+          ...MenuItem
+        }
+      }
+    }
+    footerMenu: menu(handle: $footerMenuHandle) {
+      id
+      items {
+        ...MenuItem
+        items {
+          ...MenuItem
+        }
+      }
+    }
+  }
+`;
